requirements.txt
undetected-chromedriver
selenium
python-dotenv
webdriver-manager
requests
